[ORO-0] enable managed memory api, add managed memory test, fix all warnings, cleanup

+++ Test/WMMA/main.cpp
@@ -90,7 +90,7 @@ int main( int argc, char** argv )
 
 	// Compile the WMMA kernel
 	printf( "Compiling WMMA kernel...\n" );
-	rtc_e = orortcCompileProgram( prog, opts.size(), opts.data() );
+	rtc_e = orortcCompileProgram( prog, static_cast<int>(opts.size()), opts.data() );
 	if( rtc_e != ORORTC_SUCCESS )
 	{
 		size_t logSize;
